Refactor: Update page content
